test/test-requirements.txt
envassert
fabric
nose
nose-timer
teamcity-messages
python-vagrant
requests
